tidyied up channel.hpp (made client_rights private)
